changed int to long for higher score numbers to work
fixed scoring
applied brodys asset overrides

+++ Assets/Scripts/GameManager/GameData.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class GameData
 {
     public int gameLevel;
-    public int power;
-    public int money;
+    public long power;
+    public long money;
     //public List<string> generators;
